backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
psycopg[binary]==3.2.6
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib==1.7.4
pydantic[email]==2.10.6
pydantic-settings==2.7.1
python-telegram-bot==21.10
alembic==1.14.1
